Basic Servlet Context
Added code necessary to run a very basic servlet context. The context is
accessible at port 8080 and the REST endpoint is at `/s2dr`. It
currently supports a single GET request that returns a JSON object
containing the message "Hello World". The servlet is run with
`gradle appRun` from the project root.