37 - Change Document Id to be Document Name
Updated the document table in the database to use the document name as
the document id. This was done to make fulfilling the project
requirements easier (regarding a question about the scenario where the
document exists when uploading..."If you have the permission to write,
then overwrite it."). All Java code related to the document id/name was
updated to work with this change too.

Fixes #37