17 - Add Time Parameter To Document Permissions
It is now possible to set a time limit on a delegated permission. To
set this from the client, you simply add a `timeLimitMillis` parameter
to the request body which should represent the length of time the
permission should be valid in milliseconds (i.e. if you are setting a
time limit of 30 seconds, then the request body param should be 30000).

The paramater is stored as a `TIMESTAMP` field in the database and is
nullable. A null value in this column represents an unlimited time
limit.

Fixes #17