6 - Check User Possesses Permission
When a user performed an action on a document, we were just blindly
allowing the action. This commit adds the functionality where we first
check if the user has the proper permission before reading, writing (not
including original upload), or deleting.

I also cleaned up some logging so it would be easier to follow what
actions were taking place.

Fixes #6