48 - Strengthen Permission Delegation Support
As of this commit, the server now checks if a user has a permission (or
a dominating permission) before allowing the user to delegate that
permission to someone else. It also check if there is a time limit for
the delegating user, and if so caps the time limit of the delegation to
match the max of the user.

Fixes #48