11 Document Signature/Verification Support
Added support for the `INTEGRITY` security flag. When this is selected
for uploading a document, the client must include the RSA-SHA256
signature as a request body param, which will be stored alongside the
document in the database. When a user checks out a document containing
this security flag, the server will first verify the signature using the
RSA public key of the uploader before returning the document.
- A user's RSA modulus and RSA public exponent is now stored in the User
  table so that we can reconstruct the RSA public key when verify
  signatures.
- The document's RSA-SHA256 signature is stored alongside the database
  whenever the INTEGRITY security flag is chosen.
- Both the CONFIDENTIALITY and INTEGRITY flags can be set on an upload
  at the same time.

Fixes #11