52 Permission Delegation ALL Support
Added the ability for a user to delegate a permission to all users by
specifying "ALL" in the request. If there is a permission for "ALL", it
is only considered valid when checking if a user can read or write. It
*does not* consider "ALL" permissions when determining if a user can
propogate a permission further. In order for a user to propagate a
permission, the permission must have been given to them individually
with the propogate boolean as true.

Fixes #52