53 Permission Override Support
One of the questions answered in Piazza mentioned that a user should
have the ability to "revoke" a permission from a user. This required
that permission stored in the database (and are still valid from from a
time limit aspect) should be over-ridding when a user is re-delegated
that permission.

To make this easier, I removed the `BOTH` DocumentPermission. Now if we
want to delegate both read and write permission to a user, we must
submit `READ` and `WRITE`. The delegate permission endpoint was updated
so that the `permissions` request body parameter is now a list of
permissions so that multiple can be sent in the same request.

Fixes #53