49 - Persist Database Part 1
This commit changes the database connection provider to use a persistant
database rather than the in-memory we have been using for development.
This commit includes the database file, however a follow-up commit will
add the file to the gitignore so that database changes aren't pushed and
pulled to the repository. This effectively makes the database
machine-specific.